% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amicus.R
\docType{data}
\name{amicus}
\alias{amicus}
\title{Amicus}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48445 rows and 3 columns.
}
\usage{
amicus
}
\description{
Amicus dataset
}
\examples{
  amicus
}
\keyword{datasets}
