% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{articles}
\alias{articles}
\title{Articles}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125229 rows and 2 columns.
}
\usage{
articles
}
\description{
Articles dataset
}
\examples{
  articles
}
\keyword{datasets}
