% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{constitution}
\alias{constitution}
\title{Constitution of Colombia}
\format{
A data frame with two variables:
\describe{
\item{\code{article}}{Article number}
\item{\code{text}}{text}
\item{\code{nav}}{nav structure}
}
}
\source{
https://www.constitucioncolombia.com/
}
\usage{
constitution
}
\description{
The \code{constitution} dataset 380 articles of the Colombian Constitution.
}
\examples{
  constitution
}
\keyword{datasets}
