% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official.R
\docType{data}
\name{filed_reviews}
\alias{filed_reviews}
\title{Historical dataset of filed claims for Judicial Review}
\format{
A data frame with two variables:
\describe{
\item{\code{year}}{Year}
\item{\code{n}}{Number of claims filed}
}
}
\source{
https://www.corteconstitucional.gov.co/lacorte/estadisticas
}
\usage{
filed_reviews
}
\description{
The \code{filed_reviews} dataset contains historical information extracted from the
Court's official website.
}
\examples{
  filed_reviews
}
\keyword{datasets}
