% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official.R
\docType{data}
\name{filed_tutelas}
\alias{filed_tutelas}
\title{Historical dataset of filed Tutelas}
\format{
A data frame with three variables:
\describe{
\item{\code{year}}{Year}
\item{\code{n_country}}{Number of tutelas received by the Court}
\item{\code{n_court}}{Number of tutela rulings made by the Court}
}
}
\source{
https://www.corteconstitucional.gov.co/lacorte/estadisticas
}
\usage{
filed_tutelas
}
\description{
The \code{filed_tutelas} dataset contains historical information extracted from the
Court's official website.
}
\examples{
  filed_tutelas
}
\keyword{datasets}
