% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_collection}
\alias{get_collection}
\title{Get Collection of Rulings}
\usage{
get_collection(
  x = c("gender", "peace", "treaties", "legislative_decrees", "codes", "jctt")
)
}
\arguments{
\item{x}{collection name}
}
\value{
a list of ruling ids corresponding to that collection
}
\description{
All collections come from the Court's official website, with the exception of
\code{jctt}, which comes from the "Justicia Constitucional en Tiempos de Transición"
project.
}
\details{
https://www.corteconstitucional.gov.co/relatoria/temas-interes
http://justiciatransicional.uniandes.edu.co/web/
}
\examples{

get_collection("peace")

}
