% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\docType{data}
\name{person}
\alias{person}
\title{Person}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 131443 rows and 7 columns.
}
\usage{
person
}
\description{
Judge dataset
}
\examples{
  person
}
\keyword{datasets}
