% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_article}
\alias{print_article}
\title{Get ruling summary}
\usage{
print_article(x)
}
\arguments{
\item{x}{ruling id}
}
\value{
Spanish and English summaries
}
\description{
This is a convenience function that finds the summary of
a ruling in the \code{rulings} dataset and then prints it to
the console.
}
\examples{

print_article(363)

}
