% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official.R
\docType{data}
\name{reviews}
\alias{reviews}
\title{Historical dataset of filed claims for Judicial Review}
\format{
A data frame with three variables:
\describe{
\item{\code{year}}{Year}
\item{\code{files}}{Number of review files in the Court}
\item{\code{rulings}}{Number of constitutional review rulings made by the Court}
}
}
\source{
https://www.corteconstitucional.gov.co/lacorte/estadisticas
}
\usage{
reviews
}
\description{
The \code{reviews} dataset contains historical information extracted from the
Court's official website.
}
\examples{
  reviews
}
\keyword{datasets}
