% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{rj_citations}
\alias{rj_citations}
\title{Citations (Cosa Juzgada)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4810 rows and 2 columns.
}
\usage{
rj_citations
}
\description{
RJ articles dataset
}
\examples{
  rj_citations
}
\keyword{datasets}
