% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ruling_summary}
\alias{ruling_summary}
\title{Get ruling summary}
\usage{
ruling_summary(x)
}
\arguments{
\item{x}{ruling id}
}
\value{
English and Spanish summaries
}
\description{
This is a convenience function that prints the summary of a ruling to the
console.
}
\examples{

ruling_summary("C-055-22")

}
