% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulings.R
\docType{data}
\name{rulings}
\alias{rulings}
\title{Rulings}
\format{
A data frame with nine variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{type}}{Type of Ruling}
\item{\code{n}}{Number of Judges}
\item{\code{rj}}{Indicator of Res Judicata (or "Cosa Juzgada")}
\item{\code{chamber}}{The kind of chamber that authored the ruling}
\item{\code{summary_en}}{Short summary in English.}
\item{\code{summary_es}}{Short summary in Spanish.}
}
}
\usage{
rulings
}
\description{
Rulings dataset
}
\examples{
  rulings
}
\keyword{datasets}
