% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_ruling}
\alias{summary_ruling}
\title{Get ruling summary}
\usage{
summary_ruling(x)
}
\arguments{
\item{x}{ruling id}
}
\value{
Spanish and English summaries
}
\description{
This is a convenience function that finds the summary of
a ruling in the \code{rulings} dataset and then prints it to
the console.
}
\examples{

summary_ruling("C-055-22")

}
