% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official.R
\docType{data}
\name{tutelas}
\alias{tutelas}
\title{Historical dataset of filed Tutelas}
\format{
A data frame with three variables:
\describe{
\item{\code{year}}{Year}
\item{\code{files}}{Number of tutelas received by the Court}
\item{\code{rulings}}{Number of tutela rulings made by the Court}
}
}
\source{
https://www.corteconstitucional.gov.co/lacorte/estadisticas
}
\usage{
tutelas
}
\description{
The \code{tutelas} dataset contains historical information extracted from the
Court's official website.
}
\examples{
  tutelas
}
\keyword{datasets}
