% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amicus.R
\docType{data}
\name{amicus}
\alias{amicus}
\title{Amicus}
\format{
A data frame with three variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{name}}{Name of the individual}
\item{\code{affiliation}}{An organization, government agency, university, NGO, union, social movement, etc.}
}
}
\usage{
amicus
}
\description{
The \code{amicus} dataset contains a list of individuals and organizations who
intervened in the ruling by presenting an "amicus curiae."
}
\examples{
  amicus
}
\keyword{datasets}
