% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\docType{data}
\name{appointed_judges}
\alias{appointed_judges}
\title{Appointed Judges}
\format{
A data frame with two variables:
\describe{
\item{\code{name}}{Name of the judge}
\item{\code{start}}{Start Date}
\item{\code{end}}{End Date}
}
}
\usage{
appointed_judges
}
\description{
The \code{appointed_judges} dataset contains information on the start and end periods
of the appointed judges. The cutoff date for all of them is April 3, 2024, which
note a real ending date but it is the moment in which I stopped collecting data.
}
\examples{
  appointed_judges
}
\keyword{datasets}
