% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{articles}
\alias{articles}
\title{Edge list of rulings and articles of the Colombian Constitution}
\format{
A data frame with two variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{article}}{Article number}
}
}
\usage{
articles
}
\description{
The \code{articles} dataset contains a list of articles from the Colombian
Constitution that were mentioned during the ruling.
}
\examples{
  articles
}
\keyword{datasets}
