% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccLLM-package.R
\docType{package}
\name{cccLLM-package}
\alias{cccLLM}
\alias{cccLLM-package}
\title{cccLLM: Colombian Constitutional Court LLM Dataset}
\description{
This package contains datasets about the Colombian Constitutional Court.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acastroaraujo/cccLLM}
  \item \url{https://acastroaraujo.github.io/cccLLM/}
  \item Report bugs at \url{https://github.com/acastroaraujo/cccLLM/issues}
}

}
\author{
\strong{Maintainer}: Andrés Castro Araújo \email{andres.castro.araujo@duke.edu}

}
\keyword{internal}
