% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccLLM-package.R
\docType{package}
\name{cccLLM-package}
\alias{cccLLM}
\alias{cccLLM-package}
\title{cccLLM: Colombian Constitutional Court LLM Dataset}
\description{
This package contains datasets about the Colombian Constitutional Court.
}
\author{
\strong{Maintainer}: andrés castro araújo \email{andres.castro.araujo@duke.edu}

}
\keyword{internal}
