% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccLLM-package.R
\docType{package}
\name{cccLLM-package}
\alias{cccLLM}
\alias{cccLLM-package}
\title{cccLLM: What the Package Does (One Line, Title Case)}
\description{
Contains datasets about the Colombian Constitutional Court
}
\author{
\strong{Maintainer}: andrés castro araújo \email{andres.castro.araujo@duke.edu}

}
\keyword{internal}
