% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\docType{data}
\name{person}
\alias{person}
\title{Person}
\format{
A data frame with two variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{name}}{Name of the judge}
\item{\code{mp}}{Reporting judge ("Magistrado Ponente")}
\item{\code{av}}{Concurring Opinion Indicator ("Aclaración de Voto")}
\item{\code{sv}}{Dissenting Opinion Indicator ("Salvamento de Voto")}
\item{\code{interim}}{Interim Judge Indicator ("Magistrado Encargado")}
\item{\code{conjuez}}{Substitute Judge Indicator ("Conjuez")}
}
}
\usage{
person
}
\description{
The \code{person} dataset contains information on the judges that signed off on a
particular ruling.
}
\examples{
  person
}
\keyword{datasets}
