% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{rj_citations}
\alias{rj_citations}
\title{Citations (Cosa Juzgada)}
\format{
A data frame with two variables:
\describe{
\item{\code{from}}{Ruling ID}
\item{\code{to}}{Ruling ID of precedent}
}
}
\usage{
rj_citations
}
\description{
The \code{rj_citations} dataset contains and edge list of rulings that referenced
a previous ruling as binding precedent for the current ruling. The Latin
word for this is "res judicata" and the Colombian Court uses the expression
"cosa juzgada"
}
\examples{
  rj_citations
}
\keyword{datasets}
