% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulings.R
\docType{data}
\name{rulings}
\alias{rulings}
\title{Data set of rulings}
\format{
A data frame with nine variables:
\describe{
\item{\code{id}}{Ruling ID}
\item{\code{date}}{Date}
\item{\code{type}}{Type of Ruling}
\item{\code{chamber}}{The kind of chamber that authored the ruling}
\item{\code{rj}}{Indicator of Res Judicata (or "Cosa Juzgada")}
\item{\code{n_person}}{Number of judges}
\item{\code{n_amicus}}{Number of Amicus interventions}
\item{\code{summary_en}}{Short summary in English.}
\item{\code{summary_es}}{Short summary in Spanish.}
}
}
\usage{
rulings
}
\description{
The \code{rulings} dataset contains high-level information on each rulings.
}
\examples{
  rulings
}
\keyword{datasets}
